Some notes on Region tracking
There are two strategies for tracking the regions.  Strict or lazy.

Guido proposed lazy where the region is only checked on certain `send` like operations.  This has worse overall complexity, but can be less invasive on the runtime.

Strict requires all the changes in topology to be tracked. This is an invasive change, but could be more efficient.