STORE_ATTR specialisation CHECK_WRITE
There are numerous versions of STORE_ATTR that can be specialised to.  These can more directly access the attribute if it is stored in a simple way.  The optimised bytecode is moved to after the unoptimised version has been initially executed, which makes testing challenging.

This adds checks to the three optimised versions.