Phase 1
This PR makes the changes to add deep immutability as a first-class language construct in Python.
Objects marked as immutable via the new `make_immutable` built-in function are immutable from
that point forward, and this property extends to all objects in their graph as well as their type.
Attempts to edit immutable objects will result in the new `NotWriteableError` being raised.

Signed-off-by: Matthew Johnson <matjoh@microsoft.com>