Make _Py_NewReference reset the region to local (#20)
* Add pooling failure test.

* Make _Py_NewReference reset the region to local

Some pooling code does not call _PyObject_Init, so but instead just calls _Py_NewReference.  This is problematic for us as it means the region is not reset on reallocating from a Pool.

It appears from inspection that _Py_NewReferenceNoTotal is not used in special cases that should keep the same region.